:art: Added crossover-related functions and features, changed size_t to int type for easy size checking, and improved all parts of the code related to size checking and backtracking fields.

+++ src/fhe_protobuf_mutator.cpp
@@ -18,7 +18,7 @@ namespace fhe_protobuf_mutator {
         of.close();
     }
 
-    size_t TestMessageHandler::TransferMessageType(const TestRootMsg& input, unsigned char **out_buf){
+    int TestMessageHandler::TransferMessageType(const TestRootMsg& input, unsigned char **out_buf){
         // of << input.param().num32() << " " << input.param().num64() << std::endl;
         std::string buffer;
         input.SerializePartialToString(&buffer);
